# src/api/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Ingredient": {
				"properties": {
					"name": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unit": {
						"type": "string"
					},
					"is_optional": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"quantity",
					"unit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CookType": {
				"type": "string",
				"enum": [
					"pressure-cooker",
					"stovetop",
					"oven",
					"slow-cooker",
					"grill"
				]
			},
			"RecipeCenterpieceType": {
				"type": "string",
				"enum": [
					"pork",
					"beef",
					"poultry",
					"vegetarian",
					"fish",
					"dessert"
				]
			},
			"Recipe": {
				"properties": {
					"name": {
						"type": "string"
					},
					"ingredients": {
						"items": {
							"$ref": "#/components/schemas/Ingredient"
						},
						"type": "array"
					},
					"instructions": {
						"type": "string"
					},
					"cookType": {
						"$ref": "#/components/schemas/CookType"
					},
					"totalTimeMinutes": {
						"type": "number",
						"format": "double"
					},
					"recipeCenterpieceType": {
						"$ref": "#/components/schemas/RecipeCenterpieceType"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"ingredients",
					"instructions",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "recipecate-api",
		"version": "2.0.0",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {
			"name": "matt@mtgibbs.xyz"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ingredients": {
			"get": {
				"operationId": "GetAllIngredients",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredients"
				],
				"security": [],
				"parameters": []
			}
		},
		"/recipes": {
			"get": {
				"operationId": "GetRecipes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Recipe"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "textSearch",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cookType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/CookType"
						}
					},
					{
						"in": "query",
						"name": "recipeCenterpieceType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/RecipeCenterpieceType"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Recipe"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Recipe"
							}
						}
					}
				}
			}
		},
		"/recipes/shopping-list": {
			"get": {
				"operationId": "GetShoppingListByIds",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Ingredient"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ids",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				]
			}
		},
		"/recipes/{id}": {
			"get": {
				"operationId": "GetRecipeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Recipe"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Recipe"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Recipe"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRecipe",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/unit-of-measure": {
			"get": {
				"operationId": "GetAllUnitOfMeasures",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Unit of Measure"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}